app/backend/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai==0.26.4
azure-identity==1.13.0b3
Flask==2.2.2
langchain==0.0.78
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
azure-ai-formrecognizer==3.2.1
pypdf==3.5.0
python-magic==0.4.27
